+++ tests/indicators/test_indicators.py
@@ -71,6 +71,12 @@ class TestIndicators(IndicatorTestBase):
         test.calculate()
         assert self.verify(test.as_list(), expected_highlowaverage)
 
+    @pytest.mark.usefixtures("candles", "expected_highlowcloseaverage")
+    def test_highlowcloseaverage(self, candles, expected_highlowcloseaverage):
+        test = indicators.HighLowCloseAverage(candles=candles)
+        test.calculate()
+        assert self.verify(test.as_list(), expected_highlowcloseaverage)
+
     @pytest.mark.usefixtures("candles", "expected_hma")
     def test_hma(self, candles, expected_hma):
         test = indicators.HMA(candles=candles)